clrm: Create an extra initrd with the CLRM-specific logic
Create a custom initrd with all that CLRM-specific logic and append
that to the "normal" initrd created by mkinitcpio.

This disentangles the CLRM-specific setup from mkinitcpio, so that
it should be easier to use any other initrd creation tool that
produces a systemd-based initrd.

The disentaglement is not perfect though: The modules need to be
included in the main initrd since that is where depmod is run.